+++ sdgym/__init__.py
@@ -17,6 +17,7 @@ from sdgym.cli.collect import collect_results
 from sdgym.cli.summary import make_summary_spreadsheet
 from sdgym.datasets import get_available_datasets, load_dataset
 from sdgym.synthesizers import create_sdv_synthesizer_variant, create_single_table_synthesizer
+from sdgym.result_explorer import ResultsExplorer
 
 # Clear the logging wrongfully configured by tensorflow/absl
 list(map(logging.root.removeHandler, logging.root.handlers))
@@ -30,4 +31,5 @@ __all__ = [
     'get_available_datasets',
     'create_sdv_synthesizer_variant',
     'create_single_table_synthesizer',
+    'ResultsExplorer',
 ]
