+++ sdgym/result_explorer/__init__.py
@@ -1,0 +1,5 @@
+"""Benchmark Results Explorer for SDGym."""
+
+from sdgym.result_explorer.result_explorer import ResultsExplorer
+
+__all__ = ['ResultsExplorer']
