--- sdgym/sdgym_result_explorer/__init__.py
@@ -1,5 +1,0 @@
-"""Benchmark Results Explorer for SDGym."""
-
-from sdgym.sdgym_result_explorer.result_explorer import SDGymResultsExplorer
-
-__all__ = ['SDGymResultsExplorer']
